% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getAlpha_mu}
\alias{getAlpha_mu}
\title{Returns the matrix of paramters alpha_mu}
\usage{
getAlpha_mu(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the matrix of alpha_mu parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of parameters associated with W for the mean part (mu)
}
\examples{
a <- zinbModel(n=5, J=10)
getAlpha_mu(a)
}
