% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getAlpha_pi}
\alias{getAlpha_pi}
\title{Returns the matrix of paramters alpha_pi}
\usage{
getAlpha_pi(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the matrix of alpha_pi parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of parameters associated with W for the zero part (pi)
}
\examples{
a <- zinbModel(n=5, J=10)
getAlpha_pi(a) 
}
