% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getBeta_pi}
\alias{getBeta_pi}
\title{Returns the matrix of paramters beta_pi}
\usage{
getBeta_pi(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the matrix of beta_pi parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of parameters associated with X_pi
}
\examples{
a <- zinbModel(n=5, J=10)
getBeta_pi(a)
}
