% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getEpsilon_gamma_mu}
\alias{getEpsilon_gamma_mu}
\title{Returns the vector of regularization parameter for gamma_mu}
\usage{
getEpsilon_gamma_mu(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the regularization parameters for \code{gamma_mu}.
}
\description{
Given an object describing a ZINB model, returns a vector of size the number
of columns in the parameter \code{gamma_mu} with the regularization
parameters associated to each row.
}
\examples{
a <- zinbModel(n=5, J=10)
getEpsilon_gamma_mu(a)
}
