% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getEpsilon_zeta}
\alias{getEpsilon_zeta}
\title{Returns the regularization parameter for the dispersion parameter}
\usage{
getEpsilon_zeta(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the regularization parameters for \code{zeta}.
}
\description{
The regularization parameter penalizes the variance of zeta, the log of 
the dispersion parameters across samples.
}
\examples{
a <- zinbModel(n=5, J=10)
getEpsilon_zeta(a) 
}
