% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getGamma_mu}
\alias{getGamma_mu}
\title{Returns the matrix of paramters gamma_mu}
\usage{
getGamma_mu(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the matrix of gamma_mu parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of parameters associated with V_mu
}
\examples{
a <- zinbModel(n=5, J=10)
getGamma_mu(a)
}
