% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getGamma_pi}
\alias{getGamma_pi}
\title{Returns the matrix of paramters gamma_pi}
\usage{
getGamma_pi(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the matrix of gamma_pi parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of parameters associated with V_pi
}
\examples{
a <- zinbModel(n=5, J=10)
getGamma_pi(a) 
}
