% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getPi}
\alias{getPi}
\title{Returns the matrix of probabilities of zero}
\usage{
getPi(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the matrix of probabilities of 0
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of probabilities of 0.
}
\details{
Note that although the user interface of \code{\link{zinbFit}}
requires a J x n matrix, internally this is stored as a n x J matrix (i.e., 
samples in row and genes in column). Hence the parameter matrix returned by
this function is of n x J dimensions.
}
\examples{
a <- zinbModel(n=5, J=10)
getPi(a) 
}
