% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getV_pi}
\alias{getV_pi}
\title{Returns the gene-level design matrix for pi}
\usage{
getV_pi(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the gene-level design matrix for pi
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the gene-level design matrix for pi
}
\examples{
a <- zinbModel(n=5, J=10)
getV_pi(a) 
}
