% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getX_pi}
\alias{getX_pi}
\title{Returns the sample-level design matrix for pi}
\usage{
getX_pi(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the sample-level design matrix for pi
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the sample-level design matrix for pi
}
\examples{
a <- zinbModel(n=5, J=10)
getX_pi(a)
}
