% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getZeta}
\alias{getZeta}
\title{Returns the vector of log of inverse dispersion parameters}
\usage{
getZeta(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the vector \code{zeta} of log of inverse dispersion parameters
}
\description{
Given an object that describes a matrix of zero-inflated negative binomial
distributions, returns the vector \code{zeta} of log of inverse dispersion
parameters
}
\examples{
a <- zinbModel(n=5, J=10)
getZeta(a)
}
