% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nFactors}
\alias{nFactors}
\title{Generic function that returns the number of latent factors}
\usage{
nFactors(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving latent
factors}
}
\value{
the number of latent factors
}
\description{
Given an object that describes a dataset or a model involving latent factors,
this function returns the number of latent factors.
}
