% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nFeatures}
\alias{nFeatures}
\title{Generic function that returns the number of features}
\usage{
nFeatures(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model.}
}
\value{
the number of features.
}
\description{
Given an object that describes a dataset or a model, it returns the number of
features.
}
