% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\name{nParams}
\alias{nParams}
\alias{nParams,ZinbModel-method}
\title{Generic function that returns the total number of parameters of the model}
\usage{
nParams(model)

\S4method{nParams}{ZinbModel}(model)
}
\arguments{
\item{model}{an object that describes a dataset or a model.}
}
\value{
the total number of parameters of the model.
}
\description{
Given an object that describes a model or a dataset, it returns total number of
parameters of the model.
}
\section{Functions}{
\itemize{
\item \code{nParams(ZinbModel)}: returns the total number of parameters in the model.

}}
\examples{
a <- zinbModel()
nParams(a)
}
