% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_methods.R
\name{toydata}
\alias{toydata}
\title{Toy dataset to check the model}
\format{A matrix of integers (counts) with 96 samples (rows) and 500 genes
  (columns).}
\description{
Toy dataset to check the model
}
