% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{zinb.AIC}
\alias{zinb.AIC}
\alias{zinb.AIC,ZinbModel,matrix-method}
\title{Compute the AIC of a model given some data}
\usage{
zinb.AIC(model, x)

\S4method{zinb.AIC}{ZinbModel,matrix}(model, x)
}
\arguments{
\item{model}{an object that describes a statistical model.}

\item{x}{an object that describes data.}
}
\value{
the AIC of the model.
}
\description{
Given a statistical model and some data, this function computes the AIC
of the model given the data, i.e., the AIC of the data under the model.
}
\section{Methods (by class)}{
\itemize{
\item \code{model = ZinbModel,x = matrix}: returns the AIC of the ZINB model.
}}

