% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{zinb.BIC}
\alias{zinb.BIC}
\title{Compute the BIC of a model given some data}
\usage{
zinb.BIC(model, x)
}
\arguments{
\item{model}{an object that describes a statistical model.}

\item{x}{an object that describes data.}
}
\value{
the BIC of the model.
}
\description{
Given a statistical model and some data, this function computes the BIC
of the model given the data, i.e., the BIC of the data under the model.
}
