% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_fit.R
\name{zinb.loglik.dispersion.gradient}
\alias{zinb.loglik.dispersion.gradient}
\title{Derivative of the log-likelihood of the zero-inflated negative binomial model
with respect to the log of the inverse dispersion parameter}
\usage{
zinb.loglik.dispersion.gradient(zeta, Y, mu, logitPi)
}
\arguments{
\item{zeta}{the log of the inverse dispersion parameters of the negative
binomial}

\item{Y}{a vector of counts}

\item{mu}{a vector of mean parameters of the negative binomial}

\item{logitPi}{a vector of the logit of the probability of the zero component}
}
\value{
the gradient of the inverse dispersion parameters.
}
\description{
Derivative of the log-likelihood of the zero-inflated negative binomial model
with respect to the log of the inverse dispersion parameter
}
\seealso{
\code{\link{zinb.loglik}}, \code{\link{zinb.loglik.dispersion}}.
}
