% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_values.R
\name{zinb.loglik.matrix}
\alias{zinb.loglik.matrix}
\title{Log-likelihood of the zero-inflated negative binomial model for each entry
in the matrix of counts}
\usage{
zinb.loglik.matrix(model, x)
}
\arguments{
\item{model}{the zinb model}

\item{x}{the matrix of counts}
}
\value{
the matrix of log-likelihood of the model.
}
\description{
Given a matrix of counts, this function computes the
log-probabilities of the counts under a zero-inflated negative binomial
(ZINB) model. For each count, the ZINB distribution is parametrized by three
parameters: the mean value and the dispersion of the negative binomial
distribution, and the probability of the zero component.
}
