% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_values.R
\name{computeDevianceResiduals}
\alias{computeDevianceResiduals}
\title{Deviance residuals of the zero-inflated negative binomial model}
\usage{
computeDevianceResiduals(model, x, ignoreW = TRUE)
}
\arguments{
\item{model}{the zinb model}

\item{x}{the matrix of counts}

\item{ignoreW}{logical, if true matrix \code{W} is ignored. Default is TRUE.}
}
\value{
the matrix of deviance residuals of the model.
}
\description{
Given a matrix of counts, this function computes the
deviance residuals under a zero-inflated negative binomial
(ZINB) model.
}
