% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{zinb.AIC}
\alias{zinb.AIC}
\title{Compute the AIC of a model given some data}
\usage{
zinb.AIC(model, x)
}
\arguments{
\item{model}{an object that describes a statistical model.}

\item{x}{an object that describes data.}
}
\value{
the AIC of the model.
}
\description{
Given a statistical model and some data, this function computes the AIC
of the model given the data, i.e., the AIC of the data under the model.
}
