% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper_methods.R
\docType{methods}
\name{zinbBIC}
\alias{zinbBIC}
\alias{zinbBIC,ZinbModel,matrix-method}
\title{Compute the BIC of a model given some data}
\usage{
zinbBIC(model, x)

\S4method{zinbBIC}{ZinbModel,matrix}(model, x)
}
\arguments{
\item{model}{an object that describes a statistical model.}

\item{x}{an object that describes data.}
}
\value{
the BIC of the model.
}
\description{
Given a statistical model and some data, this function computes the BIC
of the model given the data, i.e., the BIC of the data under the model.
}
\section{Methods (by class)}{
\itemize{
\item \code{model = ZinbModel,x = matrix}: returns the BIC of the ZINB model.
}}

